STYLE:  Syncronized with TEEM base source code.  This cleans up a few more compiler warnings.

+++ Utilities/NrrdIO/sane.c
@@ -95,7 +95,7 @@ airSanity(void) {
   ninfF = ninf;
   airFPValToParts_f(&sign, &exp, &mant, nanF);
   mant >>= 22;
-  if (AIR_QNANHIBIT != mant) {
+  if (AIR_QNANHIBIT != (int)mant) {
     return airInsane_QNaNHiBit;
   }
   if (!(airFP_QNAN == airFPClass_f(nanF)
